Migrate Realtime Presence tests
